WIP on fixing nix
