+++ cmd/Cargo.toml
@@ -9,6 +9,7 @@ holochain_core_types = { path = "../core_types" }
 holochain_core = { path = "../core" }
 holochain_cas_implementations = { path = "../cas_implementations" }
 holochain_container_api = { path = "../container_api" }
+holochain_wasm_utils = { path = "../wasm_utils" }
 structopt = "0.2"
 failure = "^0.1"
 serde = "1.0"
