+++ container_api/Cargo.toml
@@ -21,6 +21,7 @@ petgraph = "0.4.13"
 
 [dev-dependencies]
 test_utils = { path = "../test_utils"}
+holochain_wasm_utils = { path = "../wasm_utils" }
 clap = "2"
 structopt = "0.2"
 tempfile = "3"
