+++ wasm_utils/src/lib.rs
@@ -17,3 +17,14 @@ pub mod api_serialization;
 pub mod macros;
 pub mod memory_allocation;
 pub mod memory_serialization;
+
+pub fn wasm_target_dir(fallback: &str) -> String {
+    match std::env::var("HC_TARGET_PREFIX") {
+        Ok(target_prefix) => {
+            let test_path = std::env::var("TEST_PATH").unwrap_or(String::new());
+            let wasm_path = std::env::var("WASM_PATH").unwrap_or(String::new());
+            format!("{}{}{}target", target_prefix, test_path, wasm_path)
+        },
+        Err(_) => fallback.to_string(),
+    }
+}
